Split backend declaration and implementation
