+++ capio/common/logger.hpp
@@ -11,6 +11,7 @@
 #include <string>
 #include <sys/mman.h>
 #include <unistd.h>
+#include <memory>
 
 #include "constants.hpp"
 #include "syscall.hpp"
