+++ capio/server/include/remote/backend.hpp
@@ -1,8 +1,10 @@
 #ifndef CAPIO_SERVER_REMOTE_BACKEND_HPP
 #define CAPIO_SERVER_REMOTE_BACKEND_HPP
-#include "common/logger.hpp"
 #include <charconv>
 #include <set>
+
+#include "common/logger.hpp"
+
 
 class RemoteRequest {
     char *_buf_recv;
