+++ capio/tests/unit/server/CMakeLists.txt
@@ -2,6 +2,7 @@
 # Target information
 #####################################
 set(TARGET_NAME capio_server_unit_tests)
+find_package(MPI REQUIRED)
 
 FetchContent_MakeAvailable(capio_cl)
 
@@ -36,7 +37,7 @@ target_include_directories(${TARGET_NAME} PRIVATE
 #####################################
 # Link libraries
 #####################################
-target_link_libraries(${TARGET_NAME} PRIVATE GTest::gtest_main rt libcapio_cl)
+target_link_libraries(${TARGET_NAME} PRIVATE GTest::gtest_main rt libcapio_cl MPI::MPI_CXX)
 
 #####################################
 # Configure tests
